refactor: Modifica padrão do controller de autorização, ficando semelhante aos demais.
