+++ src/main/java/tech/ada/java/agendamentoconsultas/controllers/AuthController.java
@@ -1,19 +1,19 @@
 package tech.ada.java.agendamentoconsultas.controllers;
 
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
-import tech.ada.java.agendamentoconsultas.model.Patient;
 import tech.ada.java.agendamentoconsultas.model.Dto.LoginRequestDto;
 import tech.ada.java.agendamentoconsultas.model.Dto.LoginResponseDto;
-import tech.ada.java.agendamentoconsultas.security.service.TokenService;
+import tech.ada.java.agendamentoconsultas.service.AuthService;
 
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 
 
 @RestController
@@ -21,16 +21,17 @@ import org.springframework.web.bind.annotation.RequestBody;
 @RequiredArgsConstructor
 public class AuthController {
 
-    private final AuthenticationManager authenticationManager;
-    private final TokenService tokenService;
+    private final AuthService authService;
 
     @PostMapping("/login")
     public LoginResponseDto login(@RequestBody @Valid LoginRequestDto request) {
-        var usenamePassword = new UsernamePasswordAuthenticationToken(request.email(),request.senha());
-        var auth = this.authenticationManager.authenticate(usenamePassword);
-
-        var token = tokenService.generatedToken((Patient) auth.getPrincipal());
-        return new LoginResponseDto(token);
+        return authService.login(request);
     }
+    
+    // @PostMapping("/logout")
+    // @ResponseStatus(HttpStatus.NO_CONTENT)
+    // public void logout() {
+    //     this.authService.logout();
+    // }
     
 }
