+++ src/main/java/tech/ada/java/agendamentoconsultas/service/AuthService.java
@@ -1,0 +1,9 @@
+package tech.ada.java.agendamentoconsultas.service;
+
+import tech.ada.java.agendamentoconsultas.model.Dto.LoginRequestDto;
+import tech.ada.java.agendamentoconsultas.model.Dto.LoginResponseDto;
+
+public interface AuthService {
+    LoginResponseDto login(LoginRequestDto request);
+    void logout();
+}
